docker/django/requirements.txt
django==3.1.7
django-rest-framework
drf-yasg==1.20.0
django-filter==2.4.0
mysqlclient==2.0.3